Modded ditamap and these three topics based on Steven's request to remove -- from topicref/@props attribute. Modded topics/title/ph so it does not come back during DocUpdate.
Did not mod navtitle, because it requests a text change in the reg to do it right or a mod to the navtitle gen scripts to convert -- to non braking space. This change would need to be made in FM, Oxygen and A.Gov Tool Navtitle gen script also. Not a good way to go. Text mod is best if approved.